backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0.1
influxdb-client>=1.43.0
python-multipart>=0.0.9
